pyproject.toml
[tool.poetry]
name = "bac-detector"
version = "0.1.0"
description = "Automated Broken Access Control (BAC) Detector for authorized security testing"
authors = ["BAC Detector Contributors"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "bac_detector" },
    { include = "demo_app" },
]
keywords = ["security", "pentest", "access-control", "idor", "appsec"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = "^3.11"
typer = { version = ">=0.15.0", extras = ["all"] }
httpx = "^0.27.0"
pydantic = "^2.7.0"
pyyaml = "^6.0.1"
rich = "^13.7.1"
structlog = "^24.1.0"
# Demo app and validate-demo command — also needed for integration tests
fastapi = "^0.111.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.7"
pytest-httpx = "^0.30.0"
ruff = "^0.4.9"
mypy = "^1.10.0"
# HTML crawl dep — only needed for Phase 7
beautifulsoup4 = "^4.12.3"

[tool.poetry.scripts]
bacdet = "bac_detector.cli.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests that start a live server (deselect with -m 'not integration')",
    "unit: marks pure unit tests with no I/O (select with -m unit)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
